pyproject.toml
[project]
name = "ai-travel-companion"
version = "0.1.0"
description = "AI Travel Companion: An AI powered travel assistant"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "pydantic==2.4.2",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "python-multipart==0.0.6",
]

[tool.run]
scripts = {run = "uvicorn main:app --reload"}